Migrate picking code to its own module
