+++ pyvista/plotting/plotting.py
@@ -21,6 +21,7 @@ from pyvista.utilities import (convert_array, convert_string_array, get_array,
 from .colors import get_cmap_safe
 from .export_vtkjs import export_plotter_vtkjs
 from .mapper import make_mapper
+from .picking import PickingHelper
 from .theme import *
 from .tools import *
 
@@ -40,7 +41,7 @@ log.setLevel('CRITICAL')
 
 
 
-class BasePlotter(object):
+class BasePlotter(PickingHelper):
     """
     To be used by the Plotter and QtInteractor classes.
 
@@ -3147,276 +3148,6 @@ class BasePlotter(object):
         if hasattr(self, 'legend'):
             self.remove_actor(self.legend, reset_camera=False)
             self._render()
-
-    def get_pick_position(self):
-        """Get the pick position/area as x0, y0, x1, y1"""
-        return self.renderer.get_pick_position()
-
-    def enable_cell_picking(self, mesh=None, callback=None, through=True,
-                            show=True, show_message=True, style='wireframe',
-                            line_width=5, color='pink', font_size=18, **kwargs):
-        """
-        Enables picking of cells.  Press r to enable retangle based
-        selection.  Press "r" again to turn it off.  Selection will be
-        saved to self.picked_cells.
-        Uses last input mesh for input
-
-        Warning
-        -------
-        Visible cell picking (``through=False``) is known to not perfrom well
-        and produce incorrect selections on non-triangulated meshes if using
-        any grpahics card other than NVIDIA. A warning will be thrown if the
-        mesh is not purely triangles when using visible cell selection.
-
-        Parameters
-        ----------
-        mesh : vtk.UnstructuredGrid, optional
-            UnstructuredGrid grid to select cells from.  Uses last
-            input grid by default.
-        callback : function, optional
-            When input, calls this function after a selection is made.
-            The picked_cells are input as the first parameter to this function.
-        through : bool, optional
-            When True (default) the picker will select all cells through the
-            mesh. When False, the picker will select only visible cells on the
-            mesh's surface.
-        show : bool
-            Show the selection interactively
-        show_message : bool, str
-            Show the message about how to use the cell picking tool. If this
-            is a string, that will be the message shown.
-        kwargs : optional
-            All remaining keyword arguments are used to control how the
-            selection is intereactively displayed
-        """
-        if hasattr(self, 'notebook') and self.notebook:
-            raise AssertionError('Cell picking not available in notebook plotting')
-        if mesh is None:
-            if not hasattr(self, 'mesh'):
-                raise Exception('Input a mesh into the Plotter class first or '
-                                + 'or set it in this function')
-            mesh = self.mesh
-
-
-        def end_pick_helper(picker, event_id):
-            if show:
-                # Use try incase selection is empty
-                try:
-                    self.add_mesh(self.picked_cells, name='_cell_picking_selection',
-                        style=style, color=color, line_width=line_width,
-                        pickable=False, **kwargs)
-                except RuntimeError:
-                    pass
-
-            if callback is not None and self.picked_cells.n_cells > 0:
-                callback(self.picked_cells)
-
-            # TODO: Deactivate selection tool
-            return
-
-
-        def through_pick_call_back(picker, event_id):
-            extract = vtk.vtkExtractGeometry()
-            mesh.cell_arrays['orig_extract_id'] = np.arange(mesh.n_cells)
-            extract.SetInputData(mesh)
-            extract.SetImplicitFunction(picker.GetFrustum())
-            extract.Update()
-            self.picked_cells = pyvista.wrap(extract.GetOutput())
-            return end_pick_helper(picker, event_id)
-
-
-        def visible_pick_call_back(picker, event_id):
-            x0,y0,x1,y1 = self.get_pick_position()
-            selector = vtk.vtkOpenGLHardwareSelector()
-            selector.SetFieldAssociation(vtk.vtkDataObject.FIELD_ASSOCIATION_CELLS)
-            selector.SetRenderer(self.renderer)
-            selector.SetArea(x0,y0,x1,y1)
-            cellids = selector.Select().GetNode(0)
-            if cellids is None:
-                # No selection
-                return
-            selection = vtk.vtkSelection()
-            selection.AddNode(cellids)
-            extract = vtk.vtkExtractSelectedIds()
-            extract.SetInputData(0, mesh)
-            extract.SetInputData(1, selection)
-            extract.Update()
-            self.picked_cells = pyvista.wrap(extract.GetOutput())
-            return end_pick_helper(picker, event_id)
-
-
-        area_picker = vtk.vtkRenderedAreaPicker()
-        if through:
-            area_picker.AddObserver(vtk.vtkCommand.EndPickEvent, through_pick_call_back)
-        else:
-            # check if mesh is triangulated or not
-            # Reference:
-            #     https://github.com/pyvista/pyvista/issues/277
-            #     https://github.com/pyvista/pyvista/pull/281
-            message = "Surface picking non-triangulated meshes is known to "\
-                      "not work properly with non-NVIDIA GPUs. Please "\
-                      "consider triangulating your mesh:\n"\
-                      "\t`.extract_geometry().tri_filter()`"
-            if (not isinstance(mesh, pyvista.PolyData) or
-                    mesh.faces.size % 4 or
-                    not np.all(mesh.faces.reshape(-1, 4)[:,0] == 3)):
-                logging.warning(message)
-            area_picker.AddObserver(vtk.vtkCommand.EndPickEvent, visible_pick_call_back)
-
-        self.enable_rubber_band_style()
-        self.iren.SetPicker(area_picker)
-
-        # Now add text about cell-selection
-        if show_message:
-            if show_message == True:
-                show_message = "Press R to toggle selection tool"
-            self.add_text(str(show_message), font_size=font_size, name='_cell_picking_message')
-        return
-
-
-    def enable_point_picking(self, callback=None, show_message=True,
-                            font_size=18, color='pink', point_size=10,
-                            use_mesh=False, show_point=True, **kwargs):
-        """Enable picking a point at the mouse location in the render view
-        using the ``P`` key. This point is saved to the ``.picked_point``
-        attrbute on the plotter. Pass a callback function that takes that
-        point as an argument. The picked point can either be a point on the
-        first intersecting mesh, or a point in the 3D window.
-
-        If ``use_mesh`` is True, the callback function will be passed a pointer
-        to the picked mesh and the point ID of the selcted mesh.
-        """
-        if hasattr(self, 'notebook') and self.notebook:
-            raise AssertionError('Point picking not available in notebook plotting')
-
-        def _end_pick_event(picker, event):
-            self.picked_point = np.array(picker.GetPickPosition())
-            self.picked_mesh = picker.GetDataSet()
-            self.picked_point_id = picker.GetPointId()
-            if show_point:
-                self.add_mesh(self.picked_point, color=color,
-                              point_size=point_size, name='_picked_point',
-                              pickable=False)
-            if hasattr(callback, '__call__'):
-                if use_mesh:
-                    callback(self.picked_mesh, self.picked_point_id)
-                else:
-                    callback(self.picked_point)
-
-        point_picker = vtk.vtkPointPicker()
-        point_picker.AddObserver(vtk.vtkCommand.EndPickEvent, _end_pick_event)
-
-        self.enable_trackball_style()
-        self.iren.SetPicker(point_picker)
-
-        # Now add text about cell-selection
-        if show_message:
-            if show_message == True:
-                show_message = "Press P to pick over the mouse"
-            self.add_text(str(show_message), font_size=font_size, name='_point_picking_message')
-        return
-
-
-    def enable_path_picking(self, callback=None, show_message=True,
-                            font_size=18, color='pink', point_size=10,
-                            line_width=5):
-        """This is a conveinance method for ``enable_point_picking`` to keep
-        track of the picked points and create a line using those points.
-
-        The line is saved to the ``.picked_line`` attribute of this plotter
-        """
-        def make_line_cells(n_points):
-            # cells = np.full((n_points-1, 3), 2, dtype=np.int)
-            # cells[:, 1] = np.arange(0, n_points-1, dtype=np.int)
-            # cells[:, 2] = np.arange(1, n_points, dtype=np.int)
-            cells = np.arange(0, n_points, dtype=np.int)
-            cells = np.insert(cells, 0, n_points)
-            return cells
-
-        the_points = []
-        the_ids = []
-
-        def _the_callback(mesh, idx):
-            if mesh is None:
-                return
-            the_ids.append(idx)
-            the_points.append(mesh.points[idx])
-            self.picked_line = pyvista.PolyData(np.array(the_points))
-            self.picked_line.lines = make_line_cells(len(the_points))
-
-            self.add_mesh(self.picked_line, color=color, name='_picked_line',
-                          line_width=line_width, point_size=point_size,
-                          pickable=False)
-            if hasattr(callback, '__call__'):
-                callback(self.picked_line)
-            return
-
-        def _clear_path_event_watcher(obj, event):
-            key = self.iren.GetKeySym()
-            if key == 'c':
-                del the_points[:]
-                del the_ids[:]
-                self.remove_actor('_picked_line')
-            return
-
-        self.iren.AddObserver("KeyPressEvent", _clear_path_event_watcher)
-        if show_message == True:
-            show_message = "Press P to pick over the mouse\nPress C to clear"
-
-        return self.enable_point_picking(callback=_the_callback, use_mesh=True,
-                font_size=font_size, show_message=show_message, show_point=False)
-
-
-    def enable_geodesic_picking(self, callback=None, show_message=True,
-                            font_size=18, color='pink', point_size=10,
-                            line_width=5):
-        """This is a conveinance method for ``enable_point_picking`` to keep
-        track of the picked points and create a geodesic path using those
-        points.
-
-        The geodesic path is saved to the ``.picked_geodesic`` attribute of
-        this plotter
-        """
-        self.picked_geodesic = pyvista.PolyData()
-        self._last_picked_idx = None
-
-        def _the_callback(mesh, idx):
-            if mesh is None:
-                return
-            point = mesh.points[idx]
-            if self._last_picked_idx is None:
-                self.picked_geodesic = pyvista.PolyData(point)
-            else:
-                surface = mesh.extract_surface().triangulate()
-                locator = vtk.vtkPointLocator()
-                locator.SetDataSet(surface)
-                locator.BuildLocator()
-                start_idx = locator.FindClosestPoint(mesh.points[self._last_picked_idx])
-                end_idx = locator.FindClosestPoint(point)
-                self.picked_geodesic = self.picked_geodesic + surface.geodesic(start_idx, end_idx)
-            self._last_picked_idx = idx
-
-            self.add_mesh(self.picked_geodesic, color=color, name='_picked_line',
-                          line_width=line_width, point_size=point_size,
-                          pickable=False)
-            if hasattr(callback, '__call__'):
-                callback(self.picked_geodesic)
-            return
-
-        def _clear_path_event_watcher(obj, event):
-            key = self.iren.GetKeySym()
-            if key == 'c':
-                self.picked_geodesic = pyvista.PolyData()
-                self.remove_actor('_picked_line')
-                self._last_picked_idx = None
-            return
-
-        self.iren.AddObserver("KeyPressEvent", _clear_path_event_watcher)
-        if show_message == True:
-            show_message = "Press P to pick over the mouse\nPress C to clear"
-
-        return self.enable_point_picking(callback=_the_callback, use_mesh=True,
-                font_size=font_size, show_message=show_message, show_point=False)
 
 
     def generate_orbital_path(self, factor=3., n_points=20, viewup=None, shift=0.0):
